Revamping language lists
Redoing language lists so that clicking on a "... most languages ..." tag expands the entire language list.

Also rewriting so that the reason why a language satisfied a query is shown in the list -- i.e. If we search for languages with at least 2 fricatives, show all the fricatives in each matching language.