Working IPA vowels, consonants, metaclasses
The first wave of plug'n'play modularity changes is done. Virtually all of the HTML is now automatically generated.

The IPA consonant and vowel tables are now up and running, but the vowel  table still lacks the polished final version (i.e. the trapezoid)

Metaclasses are now also online, and more can be added quite easily.