Revamping csvtojson process
The process is now parameterizable by constants and parameters defined in const.py.

Changed the organization of data stored in selectors.py and const.py to hopefully make things clearer, more sensible, and easier to modify.

Additionally, there are now headers in the language results pane, and the language results are no longer directly clickable to expand the language list. Instead, there is an explicit button to show the matching languages.