Restoring S19 data for use in F20 semester
Change the default dataset to match the one used in S19.

Changed the UI to re-enable selectors that were disabled in F19: complex consonants, tone, stress, word order, headedness, case, agreement.

Changed requirements.txt to avoid an unsupported upgrade to TinyDB 4 for the time being. (This might be revisited in a 3.0 release of LingDB).