Totally rewrite survey processing pipeline
This commit completely rewrites csvtojson.py as a new file csv_to_json.py,
now with 900% more underscores in the name.

The new file is more object oriented, defining a LanguageData class
to represent all data about a particular language collected by a particular
student, and a Dataset class representing all LanguageData's collected by
all students for a given semester.

The new processing logic should behave identically (or close to it, anyway)
to the old one, with the added benefits that now it:
- is better documented
- uses type annotations
- is better organized (clear divisions between functional units)
- is more Pythonic (fewer janky for loops)
- has sane imports (no more import *)
- has proper logging (no more DEBUG var)
- has proper argument parsing (no more sys.argv)
- has clearer control flow (fewer crazy helpers, more sensible helpers)
- uses snake_case naming conventions so my eyes bleed less
- reads less like a freshman summer project in Java
- is gluten-free

Testing on this project has always been a bit sparse, but I ran the two
new scripts against the old datasets and didn't notice any changes beyond
the addition of the /af/ vowel phoneme, which I believe was added sometime
in the (comparatively) recent past.

The changes to json_to_db.py are less dramatic and mostly just fix the
argument parsing so that the parsers of csv_to_json and json_to_db will
play nicely with one another.

As before, the scripts can be invoked together as:
`python -m data`
but they can also be invoked by setting PYTHONPATH=. and running:
`python data/csv_to_json.py` directly