Add F21 grammar data, enhance get_glyph_list processing
This commit adds the data from the grammar survey for the F21 semester.
It also adds a new SurveySpecification to be used to process the F21 CSV.

The only major notable changes from F19 are the reintroduction of the
question on phonetic properties (tone, complex consonants, etc.), and the
addition of a new question on a language's endangerment level.

The survey also adds some additional descriptive text to some of the phonemes,
so I've had to refine the logic responsible for extracting glyphs from
that question of the survey.