Replace incorrect glyph for voiced retroflex lateral approximant
Previously, the site had been using the symbol "ȴ" for this phoneme,
even though that symbol is the voiced alveolo-palatal lateral approximant.

As a side effect, the site reported that no language possessed that phoneme,
since the survey correctly provided the symbol "ɭ" to refer to it.

This commit fixes the error by replacing the symbol with the correct one
and regenerating all the datasets and databases.