Fix endangerment, consonant types for F21
The endangerment question on the F21 survey is entirely new, and so it
needed a new SurveySpecification, along with an entry in selectors to allow
the data to be displayed. Unfortunately there's no great mechanism in
place at the moment for querying whether a string is an element of a
particular list (e.g. "is the endangerment level one of 5,6,7 ?"), so as
a hack, I've made the endangerment level a list so it can leverage the
list-like querying.

This commit also adds a new revision of the SurveySpecification for the
consonant type question, since the values in the survey have changed.

T commit fixes a bug in which null entries could slip into
a language's JSON files.

Finally, this commit changes the current dataset to F21, as well as changing
the active selectors to align with those that will be used for the F21 semester.