Add second half of F21 data
This commit adds the second half of the data for the Fall 2021 semester.

It also updates the selectors available on the site to reflect the fact
that the case and headedness data are not present this semester.

Finally, it regenerates the static HTML
to reflect the new set of selectors.