Tidy const by adding enum classes & improving docs
Previously, I just shoved all of the important constants into
`const.py`, at the top-level. This kinda works, but it was really ugly.

That approach also failed to capture relationships between constants;
for example, there are many constants which all indicate a particular
type of thing, such as JSON keys, value types, or mapping relationship.

Previously I had been prefixing constants with a `X_` to indicate which
group they belonged to; e.g. `K_` for JSON keys, `D_` for parse dicts
(now renamed to `FuzzySearchTerms`), and so on.

From Python's view, these constants were unrelated, so we'd also end up with
ugly lists of constants, where we'd have to rewrite every valid constant
twice -- once to define its value, and once to register it in the list.

Python provides a much more natural mechanism for doing this -- enums.
This commit rewrites almost all of the constants in const.py so that
they now belong to an appropriate enum, grouped in with other similar
constants of the same kind. In some cases, an enum might not have been
an appropriate choice for whatever reason, and so I used a dataclass, or
just a bare class to act as a Namespace, instead.

A trade-off with the new approach is that descriptive names are way more
verbose than non-descriptive ones. Compare K_NETID with JsonKey.NETID.

To get the best of both worlds, I've aliased the descriptive names to
(nearly) the same short descriptive names as before, by giving most of
these enum classes single letter (or at least, shorter) abbreviations,
such as `K` instead of `JsonKey`, or `T` instead of `ValueType`.
This lets us type K.NETID instead of K_NETID. Not too shabby!

The P_ names stood for "Parameter", which was the old name I'd used for
SurveySpecifications. The P doesn't make a whole bunch of sense in
context anymore, but I kept it for historical reasons. I added a related
alias `Spec` for the SurveySpecification constructor, since it's more
clear than calling a random function called `P`.

This commit does its best to keep older callers up-to-date with the new
constant names, since these constants are widely used. However, legacy
files in the various "old" directories have not been updated. I don't
want to support dead code, so these files will be removed in a
subsequent commit, rather than being updated to use the new names.

This commit also changes the names of all the variables related to the
`parse_dict` / `parse_phrase` logic, which were completely inscrutable.

The `parse_phrase` function has been renamed `fuzzy_match_phrase`,
which is more accurate and better conveys the intended use.

`parse_dict` has been renamed `FuzzySearchTerm`, to tie in better with
the new name above. `fail_list` is renamed `poisoned_search_terms`,
which also might help to convey the meaning a little better.

To tie in with these new names, I've gone through and completely
rewritten the documentation for `fuzzy_match_phrase`, to explain all of
the relevant terminology, define some key terms, and provide a simple
example that illustrates intended use. The new docstring is quite
verbose (perhaps too verbose), but I thought it best to err on the side
of over-explanation, since I wrote the damn thing and even I can't
remember how it works coming back to it after many months or years.

This commit also improves code quality and style in various ways across
the project, including things like removing unused imports, adding type
annotations, and adding better error reporting on areas of the code that
broke when I tried to make changes. (If those error messages had been
there before my life would have been easier...)