Support FuzzySearchTerms in csv_to_json one_to_many
Previously, all of the one-to-many mappings I had to deal with in the
survey data were pretty straightforward, and I could get away with a
shortcut, wherein I only check to see whether each JSON key appears
in the text of the selected items for that question.

However, F22 complicated things by adding a new "unpredictable stress"
field, to parallel the "predictable stress" field, which was previously
just called "stress". It's tricky to distinguish between those two
answers, so I needed to add a more sophisticated processing mechanism
for this case, where I use the fuzzy_match_phrase infrastructure already
in place.

I leave the original functionality in place, and call it "simple mode";
that logic will still be used if a list of JsonKeys is provided, but no
`FuzzySearchTerm` is provided. If a `FuzzySearchTerm` but no `JsonKeys`
are provided, we'll use the new "advanced mode" which uses the fuzzy
matching logic.

If neither or both arguments are provided, an error is raised, since
we wouldn't know which case to handle in that case.

For backwards compatibility, I kept the old "stress" JSON field around,
but I changed its meaning. Now it will indicate "either predictable or
unpredictable stress". I also added a new field to old survey
specifications to track "predictable stress", so the behavior for past
semesters should not change.

Since these changes are expected to alter the JSON data from past
semesters, all data will need to be regenerated.