Regenerate past datasets to include predictable stress
As mentioned in the previous commit message, prior to F22 there was a
field named "stress" which really meant "predictable stress". The
ambiguity was never a problem until F22, when a new survey option for
"unpredictable stress" was also added.

The previous commit adds some logic which introduces a "predictable
stress" field in past years. Prior to F22, it should always equal the
"stress" field.

In F22 and later, "stress" is true when either "predictable" or
"unpredictable" stress is true.

This commit regenerates all of the historical datasets to account for
this new field