Update frontend for F22
This commit updates the frontend for use in the Fall 22 semester by
adding new selectors for the updated list of consonant types, and for
the new kinds of stress.

I've decided to add stress under a single VERY HACKY selector styled
after some freakish combination of the old STRESS selector and the old
FORMATION selector.

To accommodate that change, I've added new logic to the Bool Query
handler, which will now allow the property to be a dict, and the value
to be a str. If those arguments are provided, we'll first look `value`
up in `property`, and then use the result as the new `property` to be
queried. In this case, we assume a `value` of `True`, since we're
dealing with bools, and we're trying to check if the bool is set.

While I was in there, I also fixed up some horrible horrible code I
wrote four years ago, apparently when I wasn't aware `isinstance` was a
thing.

This commit also updates the current semester to be F22 so that the new
data will be used.