Update query selectors for F22 survey 2
The second survey for F22 added one totally new survey question (functional morphology).
and also split one existing question into two new parts (word formation frequency).

Each of these changes required changes to survey data processing, which came in an earlier
commit, as well as changes to the query selectors, which is done in this commit.

This commit adds three new query selectors, each of which is almost entirely a direct
copy-paste of an existing selector, with minimal changes:
1&2) The AFFIXATION_FREQ selector and NON_AFFIXATION_FREQ selector are both
  almost identical to the preexisting WORD_FORMATION_FREQ selector.
  It makes sense that the two are nearly identical since the two new selectors
  basically just split the old selector in half, with each one receiving roughly
  half the options (plus one or two new ones).
  (Previously this selector was named just FORMATION, which made no sense, so I
   also renamed the old selector as part of this commit)is almost identical to the preexisting
  WORD_FORMATION_FREQ selector.

3) The FUNCTIONAL_MORPHOLOGY selector is a near-copy-paste of the existing
  WORD_FORMATION selector, since both of them use "at least K of [...]" lbox style queries.

This commit also renames a field in the survey data to make results read more fluently:
  "Few languages (2 / 19) use uses little or no affixation strategies to form words"
  ==>
  "Few languages (2 / 19) use little or no affixation strategies to form words"