Tweak names of F22 word formation options
This commit makes some minor tweaks to the naming of F22 word formation options:
- The "prefixation vs suffixation frequency" selector is now just "affixation frequency"
- Several "word formation" options have changed:
    "equal affixation and other" -> "equal affixal and non-affixal"
    "no non-affixal" -> "exclusively affixal"
    "mostly isolating" -> "little/no complex word formation"

These changes are purely cosmetic in the current semester.

A minor side effect is that the values are a little less backwards compatible with previous
semesters, where the names had the prior values.

However. since we haven't really got a solid interface for querying past semesters' data
anyway at this point, this is a problem we can sort out if/when we add that functionality later.