Sketch random v3 backend thoughts
This commit lays out some of my random thoughts for a v3 rework of the backend.

The current idea is to base the backend around the idea of a Query object,
which is composed of multiple arbitrary Transformation objects.
Each Transformation maps from an input type to an output type by performing some
operation, like extracting a field from a Language, intersecting two
collections, checking the truth of a bool, or returning the length.

During a Transformation chain, clients may request that context be extracted
at arbitrary points along the way.

This context will be used as the "rationale" to indicate why a result was included.

Query objects can be used to filter large LanguageSets into smaller ones,
but they can also be used to extract arbitrary data from a LanguageSet,
for example to be used to generate a graph.

This commit isn't fully functional, it just sketches my ideas.