Improve Language typing
This commit converts a Language to a Mapping[DatapointKey, DatapointValue].
It also introduces a type DatapointDict to capture the schema of each JSON file.

This commit moves some common types back into types.py so they can be shared.

This commit bumps the required Python version to 3.11,
since it introduces a dependency on StrEnum.