Define many more transformations
This commit introduces pretty much every transformation
that I could think of that might be useful for a Query.

There's one that I left out, which is the Filter Transformation.

That one will be tricky, because it'll require defining a whole bunch of named
FilterPredicate objects to mux between.

This commit includes a stub implementation for Filter,
but I'll need to plan it more carefully before finishing it.