Revise API response format
This commit revises the format of the API reply
to make it match the format spelled out in the documentation.

Specifically, it adds a "dataset" and "query_tokens" field,
and it makes "language" a value in each item in the results list,
rather than a key in a results dictionary.

This commit also renames Datasets -> Dataset,
since that name makes more sense.