Accept incorrectly-cased Get= tokens
This commit relaxes the exact-casing restriction on
Get= parameters.

Previously, a student could type something like Get=Name, which errors out,
because the key is named "name" instead of "Name".

This is especially confusing because the convenience
extractors like GetName
all make it seem like the keys should be uppercased.

This commit simply tries the lowercase version of the key if the
case-preserving version doesn't work.