HACK: Allow duplicate languages in LanguageSet
Prior to this commit, the F17 dataset couldn't be loaded, because it contained
two languages that both happened to be named Pipil.

The LanguageSet interface assumed this would never happen, and so it crashed,
because it tried to use the language name as a unique key in a dictionary.

This commit modifies LanguageSet with some extra logic to alias duplicate
languages to "alias" languages with a letter appended to the key in the
LanguageSet.

For example, you can access the first Pipil using the key Pipil A, and
the second using the key Pipil B.

There are a number of problems with this approach; for one,
it doesn't have any impact on the API or UI; it's purely a change to the
internal LanguageSet interface, which isn't really public-facing.

For now, I consider this approach a hack that makes the F17 dataset loadable
for the time being, and we can revisit this approach down the road.