Update JSON schema
This commit updates the schema of the underlying
JSON for the language data files.

Mostly, it just renames keys to avoid spaces in names.

It makes a couple other minor tweaks to key names to make them more consistent.

It also updates some of the FuzzySearchTerms which relied on
the keys having particular specific values; those search terms should now
hard-code the appropriate search strings from the survey to make them
robust to future changes.

Finally, this commit regenerates the JSON files with the new schema.

In the future, I'll probably need to rewrite the survey processing code
(yes, again) to bring it up to speed with the new standards in the
the lingdb package.

Part of this might be relatively easy -- as a phase 1, I can probably
just swap out `JsonKey` for `DatapointKey` and leave
the majority of the rest of the logic alone.